+++ Runtime/AuthenticationServiceInternal.cs
@@ -13,6 +13,7 @@ namespace Unity.Services.Authentication
     {
         const string k_ProfileRegex = @"^[a-zA-Z0-9_-]{1,30}$";
         const string k_IdProviderNameRegex = @"^oidc-[a-z0-9-_\.]{1,15}$";
+        const string k_SteamIdentityRegex = @"^[a-zA-Z0-9]{5,30}$";
         public event Action<RequestFailedException> SignInFailed;
         public event Action SignedIn;
         public event Action SignedOut;
@@ -278,24 +279,69 @@ namespace Unity.Services.Authentication
             return UnlinkExternalTokenAsync(IdProviderKeys.Facebook);
         }
 
+        [Obsolete("This method is deprecated as of version 2.7.1. Please use the SignInWithSteamAsync method with the 'identity' parameter for better security.")]
         public Task SignInWithSteamAsync(string sessionTicket, SignInOptions options = null)
         {
-            return SignInWithExternalTokenAsync(IdProviderKeys.Steam, new SignInWithExternalTokenRequest
-            {
-                IdProvider = IdProviderKeys.Steam,
-                Token = sessionTicket,
-                SignInOnly = !options?.CreateAccount ?? false
-            });
+            return SignInWithExternalTokenAsync(IdProviderKeys.Steam,
+                new SignInWithSteamRequest
+                {
+                    IdProvider = IdProviderKeys.Steam,
+                    Token = sessionTicket,
+                    SignInOnly = !options?.CreateAccount ?? false
+                });
         }
 
+        [Obsolete("This method is deprecated as of version 2.7.1. Please use the LinkWithSteamAsync method with the 'identity' parameter for better security.")]
         public Task LinkWithSteamAsync(string sessionTicket, LinkOptions options = null)
         {
-            return LinkWithExternalTokenAsync(IdProviderKeys.Steam, new LinkWithExternalTokenRequest
+            return LinkWithExternalTokenAsync(IdProviderKeys.Steam,
+                new LinkWithSteamRequest
+                {
+                    IdProvider = IdProviderKeys.Steam,
+                    Token = sessionTicket,
+                    ForceLink = options?.ForceLink ?? false
+                });
+        }
+
+        public Task SignInWithSteamAsync(string sessionTicket, string identity, SignInOptions options = null)
+        {
+            ValidateSteamIdentity(identity);
+
+            return SignInWithExternalTokenAsync(IdProviderKeys.Steam,
+                new SignInWithSteamRequest
+                {
+                    IdProvider = IdProviderKeys.Steam,
+                    Token = sessionTicket,
+                    SteamConfig = new SteamConfig() {identity = identity},
+                    SignInOnly = !options?.CreateAccount ?? false
+                });
+        }
+
+        public Task LinkWithSteamAsync(string sessionTicket, string identity, LinkOptions options = null)
+        {
+            ValidateSteamIdentity(identity);
+
+            return LinkWithExternalTokenAsync(IdProviderKeys.Steam,
+                new LinkWithSteamRequest
+                {
+                    IdProvider = IdProviderKeys.Steam,
+                    Token = sessionTicket,
+                    SteamConfig = new SteamConfig() {identity = identity},
+                    ForceLink = options?.ForceLink ?? false
+                });
+        }
+
+        void ValidateSteamIdentity(string identity)
+        {
+            if (string.IsNullOrEmpty(identity))
             {
-                IdProvider = IdProviderKeys.Steam,
-                Token = sessionTicket,
-                ForceLink = options?.ForceLink ?? false
-            });
+                throw ExceptionHandler.BuildUnknownException("Identity cannot be null or empty.");
+            }
+
+            if (!Regex.IsMatch(identity, k_SteamIdentityRegex))
+            {
+                throw ExceptionHandler.BuildUnknownException("The provided identity must only contain alphanumeric characters and be between 5 and 30 characters in length.");
+            }
         }
 
         public Task UnlinkSteamAsync()
@@ -365,6 +411,35 @@ namespace Unity.Services.Authentication
                 throw ExceptionHandler.BuildInvalidIdProviderNameException();
             }
             return UnlinkExternalTokenAsync(idProviderName);
+        }
+
+        public Task SignInWithUsernamePasswordAsync(string username, string password)
+        {
+            return SignInWithUsernamePasswordRequestAsync(BuildUsernamePasswordRequest(username, password));
+        }
+
+        public Task SignUpWithUsernamePasswordAsync(string username, string password)
+        {
+            return SignUpWithUsernamePasswordRequestAsync(BuildUsernamePasswordRequest(username, password));
+        }
+
+        public Task AddUsernamePasswordAsync(string username, string password)
+        {
+            return AddUsernamePasswordRequestAsync(BuildUsernamePasswordRequest(username, password));
+        }
+
+        public Task UpdatePasswordAsync(string currentPassword, string newPassword)
+        {
+            if (string.IsNullOrEmpty(currentPassword) || string.IsNullOrEmpty(newPassword))
+            {
+                throw ExceptionHandler.BuildInvalidCredentialsException();
+            }
+
+            return UpdatePasswordRequestAsync(new UpdatePasswordRequest
+            {
+                Password = currentPassword,
+                NewPassword = newPassword
+            });
         }
 
         public Task SignInWithUnityAsync(string token, SignInOptions options = null)
@@ -618,6 +693,69 @@ namespace Unity.Services.Authentication
             }
         }
 
+        internal Task SignInWithUsernamePasswordRequestAsync(UsernamePasswordRequest request, bool enableRefresh = true)
+        {
+            if (State == AuthenticationState.SignedOut || State == AuthenticationState.Expired)
+            {
+                return HandleSignInRequestAsync(() => NetworkClient.SignInWithUsernamePasswordAsync(request), enableRefresh);
+            }
+
+            var exception = ExceptionHandler.BuildClientInvalidStateException(State);
+            SendSignInFailedEvent(exception, false);
+            return Task.FromException(exception);
+        }
+
+        internal Task SignUpWithUsernamePasswordRequestAsync(UsernamePasswordRequest request, bool enableRefresh = true)
+        {
+            if (State == AuthenticationState.SignedOut || State == AuthenticationState.Expired)
+            {
+                return HandleSignInRequestAsync(() => NetworkClient.SignUpWithUsernamePasswordAsync(request), enableRefresh);
+            }
+
+            var exception = ExceptionHandler.BuildClientInvalidStateException(State);
+            SendSignInFailedEvent(exception, false);
+            return Task.FromException(exception);
+        }
+
+        internal async Task AddUsernamePasswordRequestAsync(UsernamePasswordRequest request)
+        {
+            if (IsAuthorized)
+            {
+                try
+                {
+                    var response = await NetworkClient.AddUsernamePasswordAsync(request);
+                    PlayerInfo.Username = response.User?.Username;
+                    return;
+                }
+                catch (WebRequestException e)
+                {
+                    throw ExceptionHandler.ConvertException(e);
+                }
+                catch (Exception e)
+                {
+                    throw ExceptionHandler.BuildUnknownException(e.Message);
+                }
+            }
+            else
+            {
+                throw ExceptionHandler.BuildClientInvalidStateException(State);
+            }
+        }
+
+        internal Task UpdatePasswordRequestAsync(UpdatePasswordRequest request, bool enableRefresh = true)
+        {
+            if (IsAuthorized)
+            {
+                // Player is signed in, update the credentials (sessionToken, accessToken)
+                return HandleUpdatePasswordRequestAsync(() => NetworkClient.UpdatePasswordAsync(request), enableRefresh);
+            }
+            else
+            {
+                var exception = ExceptionHandler.BuildClientInvalidStateException(State);
+                return Task.FromException(exception);
+            }
+        }
+
         internal Task RefreshAccessTokenAsync()
         {
             if (IsSignedIn)
@@ -663,6 +801,32 @@ namespace Unity.Services.Authentication
                 }
 
                 SendSignInFailedEvent(authException, true);
+                throw authException;
+            }
+        }
+
+        internal async Task HandleUpdatePasswordRequestAsync(Func<Task<SignInResponse>> updatePasswordRequest, bool enableRefresh = true)
+        {
+            try
+            {
+                CompleteSignIn(await updatePasswordRequest(), enableRefresh);
+            }
+            catch (RequestFailedException e)
+            {
+                SendUpdatePasswordFailedEvent(e, false);
+                throw;
+            }
+            catch (WebRequestException e)
+            {
+                var authException = ExceptionHandler.ConvertException(e);
+
+                if (authException.ErrorCode == AuthenticationErrorCodes.InvalidSessionToken)
+                {
+                    SessionTokenComponent.Clear();
+                    Logger.Log($"The session token is invalid and has been cleared. The associated account is no longer accessible through this login method.");
+                }
+
+                SendUpdatePasswordFailedEvent(authException, false);
                 throw authException;
             }
         }
@@ -874,5 +1038,24 @@ namespace Unity.Services.Authentication
             return !string.IsNullOrEmpty(idProviderName) && Regex.Match(idProviderName, k_IdProviderNameRegex).Success;
         }
 
+        UsernamePasswordRequest BuildUsernamePasswordRequest(string username, string password)
+        {
+            if (!ValidateCredentials(username, password))
+            {
+                throw ExceptionHandler.BuildInvalidCredentialsException();
+            }
+
+            return new UsernamePasswordRequest
+            {
+                Username = username,
+                Password = password
+            };
+        }
+
+        bool ValidateCredentials(string username, string password)
+        {
+            return !string.IsNullOrEmpty(username)
+                && !string.IsNullOrEmpty(password);
+        }
     }
 }