+++ Runtime/Utilities/IdProviderKeys.cs
@@ -11,6 +11,7 @@ namespace Unity.Services.Authentication
         internal const string Oculus = "oculus";
         internal const string OpenIDConnect = "oidc-";
         internal const string Steam = "steampowered.com";
+        internal const string UsernamePassword = "usernamepassword";
         internal const string Unity = "unity";
     }
 }