+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [2.7.1] - 2023-07-19
+### Added
+- Added `SignInWithSteamAsync` and `LinkWithSteamAsync` methods with `identity` parameter for better security.
+- Added `SignInWithUsernamePasswordAsync`, `SignUpWithUsernamePasswordAsync`, `AddUsernamePasswordAsync`, `UpdatePasswordAsync`
+### Changed
+- Marked previous versions of `SignInWithSteamAsync` and `LinkWithSteamAsync` methods as obsolete.
+
 ## [2.6.1] - 2023-05-31
 ### Added
 - Added `SignInWithUnityAsync` public API.